tools/auditor_tools.py
import os
import requests
from typing import Dict, Any, Optional

class FirecrawlExplorer:
    """Enterprise-grade discovery tool for mapping internal/external portals."""
    
    def __init__(self, api_key: Optional[str] = None):
        self.api_key = api_key or os.getenv("FIRECRAWL_API_KEY")
        self.base_url = "https://api.firecrawl.dev/v0" # Placeholder for Firecrawl endpoint

    def scrape_and_map(self, url: str) -> Dict[str, Any]:
        """Scrapes a portal and returns structural data for processing."""
        print(f"Firecrawl: Analyzing structural hierarchy of {url}...")
        # In a real scenario, this would call Firecrawl's /scrape or /crawl endpoint
        return {
            "url": url,
            "status": "mapped",
            "detected_fields": ["username", "password", "submit_button", "invoice_table"],
            "hierarchy": "Login > Dashboard > Billing"
        }

if __name__ == "__main__":
    explorer = FirecrawlExplorer()
    print(explorer.scrape_and_map("https://xero.com/login"))
