requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.1
python-multipart>=0.0.6
python-dotenv==1.0.1

# Database (PostgreSQL)
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Redis (task queues, pub/sub, hot cache)
redis==5.0.1

# HTTP Clients (async)
aiohttp==3.9.3
httpx>=0.26.0

# Document Intelligence (PDF + Excel + Word + Image analysis)
Pillow>=10.0.0
pytesseract>=0.3.10
PyMuPDF>=1.23.0
numpy>=1.24.0
openpyxl>=3.1.0
python-docx>=1.1.0
scipy>=1.11.0
reportlab>=4.0

# Playbook Engine (API extraction + scheduling)
jmespath==1.0.1
croniter==1.4.1

# A2A Protocol (agent-to-agent communication)
a2a-sdk[http-server]>=0.3.20

# Bastion Protocol (binary wire protocol + crypto)
pynacl>=1.5.0
msgpack>=1.0.0
